requirements.txt
# httpx
httpx==0.27.0
# dotenv
python-dotenv==1.0.1
# pandas
pandas==2.2.2
# influxdb
influxdb-client==1.42.0
# bumpversion
bump2version==1.0.1
# pre-commit
pre-commit==3.7.0
